Fixed bug in debounce algorithm. STABLE
After a long press the system keeps triggering long press events when the
software reads out the status.

Fix: plus control bit was added that locks the the long press trigger
system until the user releases the button. For that, tha counter was
reduces from 12 bits to 11 bits, but this is enough because the sample
period was changed to 1ms.

We have to handle only the leds, therefore the lcd file was renamed to
output. There will be all the interface output functions.