SPI semms to work..
There is a stable spi code (so far..) that send 4 events: DOWN, UP, SHORT,
LONG presses. It uses a stack like structure.

The base of the button debounce algorithm was modified a bit. It not
stores the short or long event only the long_lock, therefore the demo app
won't work anymore.